requirements.txt
fastapiuvicorn[standard]SQLAlchemypython-dotenvpython-jose[cryptography]passlib[bcrypt]==1.7.4bcrypt==3.2.2pydanticemail-validatorpython-multipartopenai